Decouple Execution from Deserialization (by moving its call to Application) and Implementation (by flipping the dependency).
