+++ src/Execution/IAwaitableProgramPool.cs
@@ -1,5 +1,4 @@
 ﻿using System.Threading.Tasks;
-using Symbolica.Implementation;
 
 namespace Symbolica.Execution
 {
