+++ src/Implementation/State.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Symbolica.Abstraction;
+using Symbolica.Execution;
 using Symbolica.Expression;
 using Symbolica.Implementation.Memory;
 using Symbolica.Implementation.Stack;
